api/pyproject.toml
[project]
name = "pi-monitor-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "aiosqlite>=0.20",
    "apscheduler>=3.10",
    "python-dotenv>=1.0",
    "pydantic>=2.7",
]
